+++ app/src/main/java/com/example/apptask/StockDetail.kt
@@ -1,0 +1,9 @@
+package com.example.apptask
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun StockDetail() {
+    Text(text = "test")
+}
